[Silver III] Title: 주유소, Time: 1124 ms, Memory: 182580 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -15,7 +15,7 @@ public class Main {
         for(int i = 0; i < n; i++){
             price[i] = sc.nextInt();
         }
-        
+
         // 배열의 0번째 값을 기준으로 다음 인덱스의 값을 비교
         // 그중 최소를 배열에 다시 저장
         int minPrice = price[0];
@@ -24,11 +24,11 @@ public class Main {
             price[i] = minPrice;
         }
 
-        int result = 0;
+        long result = 0;
         for(int i = 0; i < n-1; i++){
-            result += roadLength[i] * price[i];
+            result += (long) roadLength[i] * price[i];
         }
-        
+
         System.out.println(result);
     }
 }
